requirements.txt
# AI Image Detector Model VIT - 필요한 패키지들
# 기본 머신러닝 및 딥러닝 패키지
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0

# Hugging Face Transformers 및 관련 패키지
transformers>=4.30.0
datasets>=2.14.5
accelerate>=0.27.0
evaluate>=0.4.0

# 이미지 처리 패키지
Pillow>=9.0.0
opencv-python>=4.5.0

# 데이터 불균형 처리
imbalanced-learn>=0.9.0

# MLflow (모델 관리)
mlflow>=2.0.0

# Hugging Face Hub (모델 업로드용)
huggingface-hub>=0.16.0

# 기타 유틸리티
tqdm>=4.64.0
pathlib2>=2.3.0

# Jupyter 노트북 (선택사항)
jupyter>=1.0.0
ipykernel>=6.0.0

# 웹 애플리케이션 (Flask)
Flask>=2.0.0
Werkzeug>=2.0.0
